feat: identify brokers via an X-Broker-Authorization header
BREAKING_CHANGE: BROKER_ID has been renamed to BROKER_TOKEN. It is no
longer included in the server-side urls, rather it must be provided in
an X-Broker-Authorization header.

A custom header was opted for to avoid collisions with existing
authorization mechanisms. Proxy-Authorization was considered, but that
should probably accompany an overhaul to the broker to make it appear
to behave more like an HTTP proxy.